+++ app/src/main/java/com/flyn/sarcopenia_project/service/BleAction.kt
@@ -1,7 +1,0 @@
-package com.flyn.sarcopenia_project.service
-
-enum class BleAction {
-
-    GATT_CONNECTED, GATT_DISCONNECTED, EMG_DATA_AVAILABLE, ACC_DATA_AVAILABLE, GYR_DATA_AVAILABLE;
-
-}