+++ app/src/main/java/com/flyn/sarcopenia_project/service/BluetoothLeService.kt
@@ -2,36 +2,21 @@ package com.flyn.sarcopenia_project.service
 
 import android.app.Service
 import android.bluetooth.*
-import android.bluetooth.le.ScanCallback
-import android.bluetooth.le.ScanResult
-import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.content.IntentFilter
+import android.os.Binder
 import android.os.IBinder
 import android.util.Log
-import com.flyn.sarcopenia_project.toHexString
+import com.flyn.sarcopenia_project.utils.ActionManager
+import com.flyn.sarcopenia_project.utils.toShortArray
 import kotlinx.coroutines.*
+import java.util.concurrent.locks.ReentrantLock
 
 class BluetoothLeService: Service(), CoroutineScope by MainScope() {
 
     companion object {
-        const val DEVICE_NAME = "DEVICE_NAME"
         const val DATA = "BLE_DATA"
         private const val TAG = "Bluetooth Le Service"
-    }
-
-    private val leScanCallback = object: ScanCallback() {
-
-        override fun onScanResult(callbackType: Int, result: ScanResult) {
-            super.onScanResult(callbackType, result)
-            if (result.device?.name == deviceName) {
-                Log.d(TAG, "Find the target device")
-                bluetoothScan(false)
-                connect(result.device!!.address)
-            }
-        }
-
     }
 
     private val gattCallback = object: BluetoothGattCallback() {
@@ -46,171 +31,130 @@ class BluetoothLeService: Service(), CoroutineScope by MainScope() {
         override fun onConnectionStateChange(gatt: BluetoothGatt, status: Int, newState: Int) {
             if (newState == BluetoothProfile.STATE_CONNECTED) {
                 gatt.requestMtu(247)
+                gatt.requestConnectionPriority(BluetoothGatt.CONNECTION_PRIORITY_HIGH)
             }
             else if (newState == BluetoothProfile.STATE_DISCONNECTED) {
-                broadcastUpdate(BleAction.GATT_DISCONNECTED)
-                bluetoothScan(true)
+                broadcastUpdate(ActionManager.GATT_DISCONNECTED)
                 Log.i(TAG, "Disconnected from GATT server.")
             }
         }
 
         override fun onServicesDiscovered(gatt: BluetoothGatt, status: Int) {
             if (status == BluetoothGatt.GATT_SUCCESS) {
-                characteristicUUIDList.clear()
-                val uuids = setOf(UUIDList.ADC_RAW.uuid, UUIDList.IMU_ACC.uuid, UUIDList.IMU_GYR.uuid)
-                gatt.services.forEach { service ->
-                    service.characteristics.forEach { characteristic ->
-                        if (uuids.contains(characteristic.uuid)) {
-                            characteristicUUIDList.add(characteristic)
-                        }
-                    }
+                characteristicSet.clear()
+                gatt.getService(UUIDList.ADC.uuid)?.let { service ->
+                    characteristicSet.add(service.getCharacteristic(UUIDList.EMG_LEFT.uuid))
+                    characteristicSet.add(service.getCharacteristic(UUIDList.EMG_RIGHT.uuid))
+                }?: run {
+                    Log.e(TAG, "ADC service not found!")
+                    gatt.disconnect()
+                    return
                 }
-                Log.d(TAG, "uuid list size: ${characteristicUUIDList.size}")
-                broadcastUpdate(BleAction.GATT_CONNECTED)
+                gatt.getService(UUIDList.IMU.uuid)?.let { service ->
+                    characteristicSet.add(service.getCharacteristic(UUIDList.IMU_ACC.uuid))
+                    characteristicSet.add(service.getCharacteristic(UUIDList.IMU_GYR.uuid))
+                }?: run {
+                    Log.e(TAG, "IMU service not found!")
+                    gatt.disconnect()
+                    return
+                }
+                Log.d(TAG, "uuid list size: ${characteristicSet.size}")
+                broadcastUpdate(ActionManager.GATT_CONNECTED)
             }
             else Log.w(TAG, "onServicesDiscovered received: $status")
         }
 
         override fun onDescriptorWrite(gatt: BluetoothGatt, descriptor: BluetoothGattDescriptor, status: Int) {
-            waitingNotification = false
+            lock.lock()
+            condition.signal()
+            lock.unlock()
         }
 
         override fun onCharacteristicChanged(gatt: BluetoothGatt, characteristic: BluetoothGattCharacteristic) {
             when (characteristic.uuid) {
-                UUIDList.ADC_RAW.uuid -> broadcastUpdate(BleAction.EMG_DATA_AVAILABLE, characteristic)
-                UUIDList.IMU_ACC.uuid -> broadcastUpdate(BleAction.ACC_DATA_AVAILABLE, characteristic)
-                UUIDList.IMU_GYR.uuid -> broadcastUpdate(BleAction.GYR_DATA_AVAILABLE, characteristic)
-            }
-            if (bluetoothAdapter == null) {
-                Log.w(TAG, "BluetoothAdapter not initialized")
-                return
-            }
-            bluetoothGatt?.readCharacteristic(characteristic)
-//            Log.d(TAG, "readCharacteristic ${characteristic.uuid}: ${characteristic.value.toHexString()}")
-        }
-
-    }
-
-    private val commandReceiver = object: BroadcastReceiver() {
-
-        override fun onReceive(context: Context, intent: Intent) {
-            val action = BleCommand.contentAction(intent.action)?: return
-            when (action) {
-                BleCommand.DEVICE_NAME_CHANGE -> {
-                    deviceName = intent.getStringExtra(DEVICE_NAME)?: return
-                    if (!isScanning && bluetoothGatt == null) {
-                        bluetoothScan(true)
-                    }
+                UUIDList.EMG_LEFT.uuid -> {
+                    val data = EmgDecoder.decode(characteristic.value)
+                    broadcastUpdate(ActionManager.EMG_LEFT_DATA_AVAILABLE, data)
                 }
-                BleCommand.NOTIFICATION_ON -> {
-                    Log.d(TAG, "Notification ON")
-                    characteristicUUIDList.forEach {
-                        enableNotification(it, true)
-                    }
+                UUIDList.EMG_RIGHT.uuid -> {
+                    val data = EmgDecoder.decode(characteristic.value)
+                    broadcastUpdate(ActionManager.EMG_RIGHT_DATA_AVAILABLE, data)
                 }
-                BleCommand.NOTIFICATION_OFF -> {
-                    Log.d(TAG, "Notification OFF")
-                    characteristicUUIDList.forEach {
-                        enableNotification(it, false)
-                    }
+                UUIDList.IMU_ACC.uuid -> {
+                    val data = characteristic.value.toShortArray()
+                    broadcastUpdate(ActionManager.ACC_DATA_AVAILABLE, data)
+                }
+                UUIDList.IMU_GYR.uuid -> {
+                    val data = characteristic.value.toShortArray()
+                    broadcastUpdate(ActionManager.GYR_DATA_AVAILABLE, data)
                 }
             }
         }
 
     }
 
-    private val bluetoothAdapter: BluetoothAdapter? by lazy {
+    private val bluetoothAdapter: BluetoothAdapter by lazy {
         (getSystemService(Context.BLUETOOTH_SERVICE) as BluetoothManager).adapter
     }
 
-    private var isScanning = false
-    private var waitingNotification = false
-    private var deviceName = "Flyn Bluetooth Device"
     private var deviceAddress = ""
+    private var lock = ReentrantLock()
+    private var condition = lock.newCondition()
     private var bluetoothGatt: BluetoothGatt? = null
-    private var characteristicUUIDList = mutableSetOf<BluetoothGattCharacteristic>()
+    private var characteristicSet = mutableSetOf<BluetoothGattCharacteristic>()
 
-    private fun bluetoothScan(enable: Boolean) {
-        val scanner = bluetoothAdapter?.bluetoothLeScanner?: run {
-            Log.w(TAG, "Bluetooth adapter can not initialize")
-            return
-        }
-        if (enable) {
-            isScanning = true
-            scanner.startScan(leScanCallback)
-        }
-        else {
-            isScanning = false
-            scanner.stopScan(leScanCallback)
-        }
-    }
-
-    private fun connect(address: String) {
+    fun connect(address: String): Boolean {
         if (address == deviceAddress && bluetoothGatt != null) {
-            bluetoothGatt!!.connect()
             Log.d(TAG, "BLE reconnect")
-            return
+            return bluetoothGatt!!.connect()
         }
-        val device = bluetoothAdapter?.getRemoteDevice(address)?: return
+        val device = bluetoothAdapter.getRemoteDevice(address)?: return false
         bluetoothGatt = device.connectGatt(this, false, gattCallback)
         deviceAddress = address
         Log.d(TAG, "BLE connect")
+        return true
     }
 
-    private fun broadcastUpdate(action: BleAction, characteristic: BluetoothGattCharacteristic? = null) {
-        val intent = Intent(action.name)
-        if (characteristic != null) {
-            val data = characteristic.value
-            if (data != null && data.isNotEmpty()) {
-                intent.putExtra(DATA, data)
+    fun enableNotification(enable: Boolean) {
+        Log.i(TAG, "Notification is $enable")
+        GlobalScope.launch(Dispatchers.Default) {
+            characteristicSet.forEach {  characteristic ->
+                lock.lock()
+                characteristic.getDescriptor(UUIDList.CCC.uuid).run {
+                    value = if (enable) BluetoothGattDescriptor.ENABLE_NOTIFICATION_VALUE
+                    else BluetoothGattDescriptor.DISABLE_NOTIFICATION_VALUE
+                    bluetoothGatt?.writeDescriptor(this)
+                }
+                bluetoothGatt?.setCharacteristicNotification(characteristic, enable)
+                condition.await()
+                lock.unlock()
             }
         }
+    }
+
+    private fun broadcastUpdate(action: String, data: ShortArray? = null) {
+        val intent = Intent(action)
+        if (data != null) intent.putExtra(DATA, data)
         sendBroadcast(intent)
     }
 
-    private fun enableNotification(characteristic: BluetoothGattCharacteristic, enable: Boolean) {
-        Log.i(TAG, "Notification is $enable")
-        GlobalScope.launch {
-            if (bluetoothAdapter == null) return@launch
-            while (waitingNotification) {
-                delay(1)
-            }
-            waitingNotification = true
-            characteristic.getDescriptor(UUIDList.CCC.uuid).run {
-                value = if (enable) BluetoothGattDescriptor.ENABLE_NOTIFICATION_VALUE
-                else BluetoothGattDescriptor.DISABLE_NOTIFICATION_VALUE
-                bluetoothGatt?.writeDescriptor(this)
-            }
-            bluetoothGatt?.setCharacteristicNotification(characteristic, enable)
-        }
+    override fun onBind(p0: Intent?): IBinder {
+        return BleServiceBinder()
     }
 
-    override fun onBind(p0: Intent?): IBinder? = null
-
-    override fun onCreate() {
-        IntentFilter().run {
-            BleCommand.values().forEach {
-                addAction(it.name)
-            }
-            registerReceiver(commandReceiver, this)
-        }
-    }
-
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        bluetoothScan(true)
-        return super.onStartCommand(intent, flags, startId)
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        unregisterReceiver(commandReceiver)
-        if (isScanning) bluetoothScan(false)
-        if (bluetoothAdapter == null) return
-        characteristicUUIDList.forEach {
-            enableNotification(it, false)
-        }
+    override fun onUnbind(intent: Intent?): Boolean {
+        enableNotification(false)
         bluetoothGatt?.disconnect()
         Log.d(TAG, "BLE disconnect")
+        return super.onUnbind(intent)
+    }
+
+    inner class BleServiceBinder: Binder() {
+
+        fun getService(): BluetoothLeService {
+            return this@BluetoothLeService
+        }
+
     }
 
 }