+++ app/src/main/java/com/flyn/sarcopenia_project/utils/ExtraManager.kt
@@ -1,0 +1,8 @@
+package com.flyn.sarcopenia_project.utils
+
+object ExtraManager {
+
+    const val DEVICE_ADDRESS = "Device Address"
+    const val NEED_SAVE_FILE = "NEED_SAVE_FILE"
+
+}