+++ CHANGELOG.md
@@ -5,6 +5,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### [Changed]
+- use existing cloud database lib
+
+## [0.2.3] - 2022-10-19
+### [Changed]
+write file in file writing service
+
+## [0.2.2] - 2022-09-29
+### [Added]
+- adding sampling rate on screen
+- new activity can scanning device & connecting the selected device
+
+## [0.2.1] - 2022-09-28
+### [Changed]
+- save file function
+
+## [0.2.0] - 2022-09-22
+### [Changed]
+- display two channel EMG signal
 
 ## [0.1.0] - 2022-09-19
 ### [Added]